requirements.txt
annotated-types==0.5.0
anyio==3.7.1
async-timeout==4.0.2
click==8.1.6
crc16==0.1.1
exceptiongroup==1.1.2
fastapi==0.100.0
h11==0.14.0
idna==3.4
modbus-tk==1.1.3
ping3==4.0.4
pydantic==2.1.1
pydantic_core==2.4.0
pyserial==3.5
python-redis==0.3.0
python-snap7==1.3
redis==4.6.0
sniffio==1.3.0
starlette==0.27.0
typing_extensions==4.7.1
uvicorn==0.23.1
